algo mejorado

+++ Kernel/Drivers/keyboard.c
@@ -8,7 +8,7 @@ extern uint8_t getKey();
  * Buffer --> es "circular". Si se llena, pisa lo que primero se puso.
  */
 
-static uint8_t buffer[BUFFER_SIZE];
+static uint16_t buffer[BUFFER_SIZE];
 //static uint64_t buffer_index = 0;
 static uint64_t buffer_dim = 0;
 static uint64_t buffer_current = 0;
